Chore(validations): Aligning config validations across multiple endpoints (#1215)
* chore(validations): Aligning config validations across endpoints

* fixing lint issue

* chore(validations): removing unneeded comments and tests
