+++ keel-core-test/src/main/kotlin/com/netflix/spinnaker/keel/persistence/CombinedRepositoryTests.kt
@@ -7,17 +7,11 @@ import com.netflix.spinnaker.keel.api.artifacts.ArtifactType
 import com.netflix.spinnaker.keel.api.artifacts.DockerArtifact
 import com.netflix.spinnaker.keel.api.artifacts.TagVersionStrategy.BRANCH_JOB_COMMIT_BY_JOB
 import com.netflix.spinnaker.keel.api.id
-import com.netflix.spinnaker.keel.core.api.DependsOnConstraint
-import com.netflix.spinnaker.keel.core.api.SubmittedDeliveryConfig
-import com.netflix.spinnaker.keel.core.api.SubmittedEnvironment
 import com.netflix.spinnaker.keel.core.api.SubmittedResource
 import com.netflix.spinnaker.keel.core.api.normalize
 import com.netflix.spinnaker.keel.events.ArtifactRegisteredEvent
 import com.netflix.spinnaker.keel.events.ResourceCreated
 import com.netflix.spinnaker.keel.events.ResourceUpdated
-import com.netflix.spinnaker.keel.exceptions.DuplicateArtifactReferenceException
-import com.netflix.spinnaker.keel.exceptions.DuplicateResourceIdException
-import com.netflix.spinnaker.keel.exceptions.MissingEnvironmentReferenceException
 import com.netflix.spinnaker.keel.resources.ResourceSpecIdentifier
 import com.netflix.spinnaker.keel.test.DummyResourceSpec
 import com.netflix.spinnaker.keel.test.TEST_API_V1
@@ -38,7 +32,6 @@ import strikt.assertions.hasSize
 import strikt.assertions.isA
 import strikt.assertions.isEmpty
 import strikt.assertions.isEqualTo
-import strikt.assertions.isFailure
 import strikt.assertions.isSuccess
 import strikt.assertions.isTrue
 
@@ -300,148 +293,6 @@ abstract class CombinedRepositoryTests<D : DeliveryConfigRepository, R : Resourc
     }
 
     context("persisting delivery config manifests") {
-
-      context("a delivery config with non-unique resource ids errors while persisting") {
-        val submittedConfig = SubmittedDeliveryConfig(
-          name = configName,
-          application = "keel",
-          serviceAccount = "keel@spinnaker",
-          artifacts = setOf(artifact),
-          environments = setOf(
-            SubmittedEnvironment(
-              name = "test",
-              resources = setOf(
-                SubmittedResource(
-                  kind = TEST_API_V1.qualify("whatever"),
-                  spec = DummyResourceSpec("test", "im a twin", "keel")
-                )
-              ),
-              constraints = emptySet()
-            ),
-            SubmittedEnvironment(
-              name = "prod",
-              resources = setOf(
-                SubmittedResource(
-                  kind = TEST_API_V1.qualify("whatever"),
-                  spec = DummyResourceSpec("test", "im a twin", "keel")
-                )
-              ),
-              constraints = emptySet()
-            )
-          )
-        )
-
-        test("an error is thrown and config is deleted") {
-          expectCatching {
-            subject.upsertDeliveryConfig(submittedConfig)
-          }.isFailure()
-            .isA<DuplicateResourceIdException>()
-
-          expectThat(subject.allResourceNames().size).isEqualTo(0)
-        }
-      }
-
-      context("a delivery config with non-unique artifact references errors while persisting") {
-        // Two different artifacts with the same reference
-        val artifacts = setOf(
-          DockerArtifact(name = "org/thing-1", deliveryConfigName = configName, reference = "thing"),
-          DockerArtifact(name = "org/thing-2", deliveryConfigName = configName, reference = "thing")
-        )
-
-        val submittedConfig = SubmittedDeliveryConfig(
-          name = configName,
-          application = "keel",
-          serviceAccount = "keel@spinnaker",
-          artifacts = artifacts,
-          environments = setOf(
-            SubmittedEnvironment(
-              name = "test",
-              resources = setOf(
-                SubmittedResource(
-                  metadata = mapOf("serviceAccount" to "keel@spinnaker"),
-                  kind = TEST_API_V1.qualify("whatever"),
-                  spec = DummyResourceSpec(data = "o hai")
-                )
-              ),
-              constraints = emptySet()
-            )
-          )
-        )
-        test("an error is thrown and config is deleted") {
-          expectCatching {
-            subject.upsertDeliveryConfig(submittedConfig)
-          }.isFailure()
-            .isA<DuplicateArtifactReferenceException>()
-
-          expectThat(subject.allResourceNames().size).isEqualTo(0)
-        }
-      }
-
-      context("a second delivery config for an app fails to persist") {
-        val submittedConfig1 = SubmittedDeliveryConfig(
-          name = configName,
-          application = "keel",
-          serviceAccount = "keel@spinnaker",
-          artifacts = setOf(DockerArtifact(name = "org/thing-1", deliveryConfigName = configName, reference = "thing")),
-          environments = setOf(
-            SubmittedEnvironment(
-              name = "test",
-              resources = setOf(
-                SubmittedResource(
-                  metadata = mapOf("serviceAccount" to "keel@spinnaker"),
-                  kind = TEST_API_V1.qualify("whatever"),
-                  spec = DummyResourceSpec(data = "o hai")
-                )
-              ),
-              constraints = emptySet()
-            )
-          )
-        )
-
-        val submittedConfig2 = submittedConfig1.copy(name = "double-trouble")
-        test("an error is thrown and config is not persisted") {
-          subject.upsertDeliveryConfig(submittedConfig1)
-          expectCatching {
-            subject.upsertDeliveryConfig(submittedConfig2)
-          }.isFailure()
-            .isA<TooManyDeliveryConfigsException>()
-
-          expectThat(subject.getDeliveryConfigForApplication("keel").name).isEqualTo(configName)
-        }
-      }
-
-      context("submitting delivery config with invalid environment name as a constraint") {
-        val submittedConfig = SubmittedDeliveryConfig(
-          name = configName,
-          application = "keel",
-          serviceAccount = "keel@spinnaker",
-          artifacts = setOf(DockerArtifact(name = "org/thing-1", deliveryConfigName = configName, reference = "thing")),
-          environments = setOf(
-            SubmittedEnvironment(
-              name = "test",
-              resources = emptySet(),
-              constraints = emptySet()
-            ),
-              SubmittedEnvironment(
-              name = "test",
-            resources = emptySet(),
-            constraints = setOf(DependsOnConstraint(environment = "notARealEnvironment"))
-          )
-          )
-        )
-
-        test("an error is thrown and config is not persisted") {
-          expectCatching {
-            subject.upsertDeliveryConfig(submittedConfig)
-          }.isFailure()
-            .isA<MissingEnvironmentReferenceException>()
-
-          expectCatching {
-            subject.getDeliveryConfig(configName)
-          }.isFailure()
-            .isA<NoSuchDeliveryConfigException>()
-        }
-      }
     }
   }
 }
