+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/persistence/CombinedRepository.kt
@@ -12,7 +12,6 @@ import com.netflix.spinnaker.keel.api.artifacts.DockerArtifact
 import com.netflix.spinnaker.keel.api.id
 import com.netflix.spinnaker.keel.constraints.ConstraintState
 import com.netflix.spinnaker.keel.core.api.ApplicationSummary
-import com.netflix.spinnaker.keel.core.api.DependsOnConstraint
 import com.netflix.spinnaker.keel.core.api.EnvironmentArtifactPin
 import com.netflix.spinnaker.keel.core.api.EnvironmentArtifactVetoes
 import com.netflix.spinnaker.keel.core.api.EnvironmentSummary
@@ -24,9 +23,6 @@ import com.netflix.spinnaker.keel.events.ApplicationEvent
 import com.netflix.spinnaker.keel.events.ArtifactRegisteredEvent
 import com.netflix.spinnaker.keel.events.ResourceEvent
 import com.netflix.spinnaker.keel.events.ResourceHistoryEvent
-import com.netflix.spinnaker.keel.exceptions.DuplicateArtifactReferenceException
-import com.netflix.spinnaker.keel.exceptions.DuplicateResourceIdException
-import com.netflix.spinnaker.keel.exceptions.MissingEnvironmentReferenceException
 import java.time.Clock
 import java.time.Duration
 import java.time.Instant
@@ -78,8 +74,6 @@ class CombinedRepository(
         )
       }
     )
-
-    validate(new)
     return upsertDeliveryConfig(new)
   }
 
@@ -191,73 +185,6 @@ class CombinedRepository(
           deliveryConfigRepository.deleteEnvironment(new.name, environment.name)
         }
       }
-  }
-
-  /**
-   * Run validation checks against delivery config to ensure:
-   *
-   * - resources have unique ids
-   * - artifacts have unique references
-   *
-   * Throws an exception if config fails any checks
-   */
-  private fun validate(config: DeliveryConfig) {
-
-    // helper function to get duplicates in a list
-    fun duplicates(ids: List<String>): List<String> =
-      ids.groupingBy { it }
-        .eachCount()
-        .filter { it.value > 1 }
-        .keys
-        .toList()
-
-    /**
-     * check: resources have unique ids
-     */
-    val resources = config.environments.map { it.resources }.flatten().map { it.id }
-    val duplicateResources = duplicates(resources)
-
-    if (duplicateResources.isNotEmpty()) {
-      val envToResources: Map<String, MutableList<String>> = config.environments
-        .map { env -> env.name to env.resources.map { it.id }.toMutableList() }.toMap()
-      val envsAndDuplicateResources = envToResources
-        .filterValues { rs: MutableList<String> ->
-          // remove all the resources we don't care about from this mapping
-          rs.removeIf { it !in duplicateResources }
-          // if there are resources left that we care about, leave it in the map
-          rs.isNotEmpty()
-        }
-      log.error("Validation failed for ${config.name}, duplicates resource ids found: $envsAndDuplicateResources")
-      throw DuplicateResourceIdException(duplicateResources, envsAndDuplicateResources)
-    }
-
-    /**
-     * check: artifacts have unique references
-     */
-    val refs = config.artifacts.map { it.reference }
-    val duplicateRefs = duplicates(refs)
-
-    if (duplicateRefs.isNotEmpty()) {
-      val duplicatesArtifactNameToRef: Map<String, String> = config.artifacts
-        .filter { duplicateRefs.contains(it.reference) }
-        .associate { art -> art.name to art.reference }
-
-      log.error("Validation failed for ${config.name}, duplicate artifact references found: $duplicatesArtifactNameToRef")
-      throw DuplicateArtifactReferenceException(duplicatesArtifactNameToRef)
-    }
-
-    /**
-     * validating depends on environments
-     */
-    config.environments.forEach { environment ->
-      environment.constraints.forEach { constraint ->
-        if (constraint is DependsOnConstraint) {
-          config.environments.find {
-            it.name == constraint.environment
-          } ?: throw MissingEnvironmentReferenceException(constraint.environment)
-        }
-      }
-    }
   }
 
   private fun Set<DeliveryArtifact>.transform(deliveryConfigName: String) =
