+++ keel-web/src/main/kotlin/com/netflix/spinnaker/keel/rest/DeliveryConfigController.kt
@@ -5,10 +5,12 @@ import com.netflix.spinnaker.keel.core.api.SubmittedDeliveryConfig
 import com.netflix.spinnaker.keel.diff.AdHocDiffer
 import com.netflix.spinnaker.keel.diff.EnvironmentDiff
 import com.netflix.spinnaker.keel.persistence.KeelRepository
+import com.netflix.spinnaker.keel.validators.DeliveryConfigValidator
 import com.netflix.spinnaker.keel.yaml.APPLICATION_YAML_VALUE
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.parameters.RequestBody as SwaggerRequestBody
 import org.slf4j.LoggerFactory
+import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -17,13 +19,15 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping(path = ["/delivery-configs"])
 class DeliveryConfigController(
   private val repository: KeelRepository,
-  private val adHocDiffer: AdHocDiffer
+  private val adHocDiffer: AdHocDiffer,
+  private val validator: DeliveryConfigValidator
 ) {
   @Operation(
     description = "Registers or updates a delivery config manifest."
@@ -41,8 +45,10 @@ class DeliveryConfigController(
       description = "The delivery config. If its `name` matches an existing delivery config the operation is an update, otherwise a new delivery config is created."
     )
     deliveryConfig: SubmittedDeliveryConfig
-  ): DeliveryConfig =
-    repository.upsertDeliveryConfig(deliveryConfig)
+  ): DeliveryConfig {
+    validator.validate(deliveryConfig)
+    return repository.upsertDeliveryConfig(deliveryConfig)
+  }
 
   @GetMapping(
     path = ["/{name}"],
@@ -75,19 +81,21 @@ class DeliveryConfigController(
     produces = [APPLICATION_JSON_VALUE, APPLICATION_YAML_VALUE]
   )
   @PreAuthorize("@authorizationSupport.hasApplicationPermission('READ', 'APPLICATION', #deliveryConfig.application)")
-  fun diff(@RequestBody deliveryConfig: SubmittedDeliveryConfig): List<EnvironmentDiff> =
-    adHocDiffer.calculate(deliveryConfig)
+  fun diff(@RequestBody deliveryConfig: SubmittedDeliveryConfig): List<EnvironmentDiff> {
+    validator.validate(deliveryConfig)
+    return adHocDiffer.calculate(deliveryConfig)
+  }
 
   @PostMapping(
     path = ["/validate"],
     consumes = [APPLICATION_JSON_VALUE, APPLICATION_YAML_VALUE],
     produces = [APPLICATION_JSON_VALUE, APPLICATION_YAML_VALUE]
   )
+  @ResponseStatus(value = HttpStatus.NO_CONTENT)
   @PreAuthorize("@authorizationSupport.hasApplicationPermission('READ', 'APPLICATION', #deliveryConfig.application)")
-  fun validate(@RequestBody deliveryConfig: SubmittedDeliveryConfig) =
-  // TODO: replace with JSON schema/OpenAPI spec validation when ready (for now, leveraging parsing error handling
-    //  in [ExceptionHandler])
-    mapOf("status" to "valid")
+  fun validate(@RequestBody deliveryConfig: SubmittedDeliveryConfig) {
+    validator.validate(deliveryConfig)
+  }
 
   private val log by lazy { LoggerFactory.getLogger(javaClass) }
 }
